Rename `nsga2` module to `nsgaii` to follow sampler naming
