+++ optuna/samplers/__init__.py
@@ -1,4 +1,4 @@
-from optuna.samplers import nsga2
+from optuna.samplers import nsgaii
 from optuna.samplers._base import BaseSampler
 from optuna.samplers._cmaes import CmaEsSampler
 from optuna.samplers._grid import GridSampler
@@ -8,7 +8,7 @@ from optuna.samplers._search_space import intersection_search_space
 from optuna.samplers._search_space import IntersectionSearchSpace
 from optuna.samplers._tpe.multi_objective_sampler import MOTPESampler
 from optuna.samplers._tpe.sampler import TPESampler
-from optuna.samplers.nsga2._sampler import NSGAIISampler
+from optuna.samplers.nsgaii._sampler import NSGAIISampler
 
 
 __all__ = [
@@ -22,5 +22,5 @@ __all__ = [
     "RandomSampler",
     "TPESampler",
     "intersection_search_space",
-    "nsga2",
+    "nsgaii",
 ]
