+++ optuna/samplers/nsgaii/_crossovers/_sbx.py
@@ -12,7 +12,7 @@ from optuna.distributions import IntLogUniformDistribution
 from optuna.distributions import IntUniformDistribution
 from optuna.distributions import LogUniformDistribution
 from optuna.distributions import UniformDistribution
-from optuna.samplers.nsga2._crossovers._base import BaseCrossover
+from optuna.samplers.nsgaii._crossovers._base import BaseCrossover
 from optuna.study import Study
