+++ optuna/samplers/nsgaii/_sampler.py
@@ -20,9 +20,9 @@ from optuna.distributions import BaseDistribution
 from optuna.samplers._base import BaseSampler
 from optuna.samplers._random import RandomSampler
 from optuna.samplers._search_space import IntersectionSearchSpace
-from optuna.samplers.nsga2._crossover import perform_crossover
-from optuna.samplers.nsga2._crossovers._base import BaseCrossover
-from optuna.samplers.nsga2._crossovers._uniform import UniformCrossover
+from optuna.samplers.nsgaii._crossover import perform_crossover
+from optuna.samplers.nsgaii._crossovers._base import BaseCrossover
+from optuna.samplers.nsgaii._crossovers._uniform import UniformCrossover
 from optuna.study import Study
 from optuna.study import StudyDirection
 from optuna.study._multi_objective import _dominates
